Factor out storage callbacks
