+++ packages/core/echo/echo-pipeline/src/automerge/index.ts
@@ -5,6 +5,7 @@
 export * from './automerge-host';
 export * from './automerge-storage-adapter';
 export * from './automerge-doc-loader';
+export * from './leveldb-storage-adapter';
 export * from './local-host-network-adapter';
 export * from './mesh-network-adapter';
 export * from './types';
