+++ packages/core/echo/indexing/src/index.ts
@@ -3,8 +3,9 @@
 //
 
 export * from './indexer';
-export * from './index-store';
 export * from './index-metadata-store';
-export * from './types';
 export * from './index-schema';
 export * from './index-service';
+export * from './index-store';
+export * from './storage-callbacks';
+export * from './types';
